ERT-1059 Reorganization SearchBox and some renaming
