+++ devel/python/python/ert_gui/ertwidgets/checklist.py
@@ -2,8 +2,7 @@ from PyQt4.QtCore import QSize, Qt
 from PyQt4.QtGui import QToolButton, QHBoxLayout, QLabel, QListWidget
 from PyQt4.QtGui import QWidget, QVBoxLayout, QListWidgetItem, QMenu, QAbstractItemView
 
-from ert_gui.ertwidgets import addHelpToWidget
-from ert_gui.widgets.search_box import SearchBox
+from ert_gui.ertwidgets import addHelpToWidget, SearchBox
 from ert_gui.widgets.util import resourceIcon
