+++ devel/python/python/ert_gui/ertwidgets/models/CMakeLists.txt
@@ -1,11 +1,11 @@
 set(PYTHON_SOURCES
     __init__.py
-    active_realizations_model.py
+    activerealizationsmodel.py
     analysismodulevariablesmodel.py
     ertmodel.py
     ertsummary.py
     selectable_list_model.py
-    value_model.py
+    valuemodel.py
 )
 
 add_python_package("python.ert_gui.ertwidgets.models" ${PYTHON_INSTALL_PREFIX}/ert_gui/ertwidgets/models "${PYTHON_SOURCES}" True)
