+++ devel/python/python/ert_gui/ertwidgets/searchbox.py
